Refetch bounties after transactions
